+++ changelog.md
@@ -1,7 +1,9 @@
 # Changelog
-## Unreleased [0.10.0]
+## Unreleased [0.9.1]
 
 ### Added
+ * **Dockerfile** highlighting via `Kadet\Highlighter\Language\Dockerfile`
+
 ### Fixed
 
 ## 2024-01-26 [0.9.0]